% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_RData.R
\name{bedToRData}
\alias{bedToRData}
\title{Bed to RData}
\usage{
bedToRData(dirFrom)
}
\arguments{
\item{dirFrom}{The path to files. By default "~/tmp/RData/Input".}
}
\value{
In case of success it returns the word "Done".
}
\description{
This function converts files from *.bed to .RData files.
}
\examples{
bedToRData(dirFrom="")

}
\author{
Alexey Solovyev
}
