% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_RData.R
\name{combineRData}
\alias{combineRData}
\title{Combines a lot of files .RData in one file}
\usage{
combineRData(dirFrom="", dirTo="")
}
\arguments{
\item{dirFrom}{The path to input files. By default "~/tmp/RData/Input".}

\item{dirTo}{The path to output file. By default "~/tmp/RData/Output".}
}
\value{
In case of success it returns the word "Done".
}
\description{
Combines a lot of files .RData in one file.
}
\examples{
bedToRData(dirFrom="", dirTo="")

}
\author{
Alexey Solovyev
}
