% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFrame_to_IRange.R
\name{DFrameToIRange}
\alias{DFrameToIRange}
\title{DataFrame to IRange}
\usage{
DFrameToIRange(arrayData)
}
\arguments{
\item{arrayData}{The data to convert. Second column of arrayData is start of IRange,
third column of arrayData is end of IRange. For exemple, Data from one of Fetches.}
}
\value{
The data in IRange format.
}
\description{
This function converts DataFrame to IRange.
}
\examples{
myIRange  <- DFrameToIRange(data.frame)

}
\author{
Alexey Solovyev
}
