% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_by_chromosome.R
\name{divideByChromosome}
\alias{divideByChromosome}
\title{Devides by chromosomes}
\usage{
divideByChromosome(dirTo)
}
\arguments{
\item{dirTo}{The path to output files. By default "~/tmp/RData/Output".}
}
\value{
In case of success it returns the word "Done".
}
\description{
Function devides given file into set of files by name of chromosome.
}
\examples{
divideByChromosome(dirTo="")

}
\author{
Alexey Solovyev
}
