% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bed_file.R
\name{downloadBedFile}
\alias{downloadBedFile}
\title{Loads file from URL to tmp}
\usage{
downloadBedFile(list of params)
}
\arguments{
\item{targetDir}{The name of the directory to download the file in, by default "~/tmp".}

\item{version}{Version of ReMap.}

\item{species}{Name of specie.}

\item{assembly}{Number of assembly.}

\item{type}{Type of data.}

\item{force}{It can be FALSE or TRUE. If FALSE (default), then no file is overwrited and the
user is given confirmation message.}
}
\value{
The path to the downloaded file.
}
\description{
Function loads bed-files from URL to the given directory.
}
\examples{
downloadBedFile(targetDir = "", version = "2018", species = "Homo sapiens",
assembly = "hg38", type = "Test peaks", force = FALSE)

}
\author{
Alexey Solovyev
}
