% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cell.R
\name{fetchCell}
\alias{fetchCell}
\title{Select cell line}
\usage{
fetchCell(arrayData, cell)
}
\arguments{
\item{arrayData}{Data or path to data, by default "~/tmp/RData/Output".}

\item{cell}{Name of line.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given cell line.
}
\examples{
myFrame <- fetchCell(arrayData = "", cell = "vcap_shctr_r1881")

}
\author{
Alexey Solovyev
}
