% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_coords_files.R
\name{fetchCoordsFiles}
\alias{fetchCoordsFiles}
\title{Search by address}
\usage{
fetchCoordsFiles(dirFrom = "", nameChromosome = "chr21", begin = "", end = "",
firstCut = "in", secondCut = "in")
}
\arguments{
\item{dirFrom}{Path to data, by default "~/tmp/RData/Output".}

\item{nameChromosome}{Name of chromosome.}

\item{begin}{Position of starting of searching on chromosome.}

\item{end}{Position of ending of searching on chromosome.}

\item{firstCut}{Can be "in" or "out". Condition of searching. "in" means that begin of MACS-peak can be before param "begin".
"out" means that begin of MACS-peak can be only after param "begin".}

\item{secondCut}{Can be "in" or "out". Condition of searching. "in" means that end of MACS-peak can be after param "end".
"out" means that end of MACS-peak can be only before param "end".}
}
\value{
Data Frame (class = "data.frame") of peaks in this address.
}
\description{
Function selects in given address.
}
\examples{
myFrame <- fetchCoordsFiles(dirFrom = "", nameChromosome = "chr21", begin = "",
end = "", firstCut = "in", secondCut = "in")

}
\author{
Alexey Solovyev
}
