% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ID.R
\name{fetchID}
\alias{fetchID}
\title{Select ID}
\usage{
fetchID(arrayData, ID)
}
\arguments{
\item{arrayData}{Data or path to data, by default "~/tmp/RData/Output".}

\item{ID}{Name of line.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given ID.
}
\examples{
myFrame <- fetchID(arrayData = "", ID = "GSE56086")

}
\author{
Alexey Solovyev
}
