% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ID_DFrame.R
\name{fetchIDDFrame}
\alias{fetchIDDFrame}
\title{Select ID}
\usage{
fetchIDDFrame(dataFrame = "", ID = "GSE56086")
}
\arguments{
\item{dataFrame}{Data frame.}

\item{ID}{Name of ID.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given ID.
}
\examples{
myFrame <- fetchIDDFrame(dataFrame = "", ID = "GSE56086")

}
\author{
Alexey Solovyev
}
