% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ID_files.R
\name{fetchIDFiles}
\alias{fetchIDFiles}
\title{Select ID}
\usage{
fetchIDFiles(dirFrom, ID)
}
\arguments{
\item{dirFrom}{Path to data, by default "~/tmp/RData/Output".}

\item{ID}{Name of ID.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given ID.
}
\examples{
myFrame <- fetchIDFiles(dirFrom = "", ID = "GSE56086")

}
\author{
Alexey Solovyev
}
