% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_TF.R
\name{fetchTF}
\alias{fetchTF}
\title{Select TF}
\usage{
fetchTF(arrayData = "", TF = "FOXA1")
}
\arguments{
\item{arrayData}{Data or path to data, by default "~/tmp/RData/Output".}

\item{TF}{Name of TF.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given TF.
}
\examples{
myFrame <- fetchTF(arrayData = "", TF = "FOXA1")

}
\author{
Alexey Solovyev
}
