% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_TF_DFrame.R
\name{fetchTFDFrame}
\alias{fetchTFDFrame}
\title{Select TF}
\usage{
fetchTFDFrame(dataFrame, TF)
}
\arguments{
\item{TF}{Name of TF.}

\item{arrayData}{Data frame.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given TF.
}
\examples{
myFrame <- fetchTFDFrame(dataFrame = "", TF = "FOXA1")

}
\author{
Alexey Solovyev
}
