% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_TF_files.R
\name{fetchTFFiles}
\alias{fetchTFFiles}
\title{Select TF}
\usage{
fetchTFFiles(dirFrom = "", TF = "FOXA1")
}
\arguments{
\item{dirFrom}{Path to data, by default "~/tmp/RData/Output".}

\item{TF}{Name of TF.}
}
\value{
Data Frame (class = "data.frame") of peaks of this line.
}
\description{
Function selects given TF.
}
\examples{
myFrame <- fetchTFFiles(dirFrom = "", TF = "FOXA1")

}
\author{
Alexey Solovyev
}
