% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_Cell_TF.R
\name{fractionCellTF}
\alias{fractionCellTF}
\title{Fraction of factors of the dominating cellular line}
\usage{
fetchCoords(list of params)
}
\arguments{
\item{arrayData}{Data or path to data, by default "~/tmp/RData/Output".}

\item{nameChromosome}{Name of chromosome.}

\item{begin}{Position of starting of searching on chromosome.}

\item{end}{Position of ending of searching on chromosome.}

\item{firstCut}{Can be "in" or "out". Condition of searching. "in" means that begin of MACS-peak can be before param "begin".
"out" means that begin of MACS-peak can be only after param "begin".}

\item{secondCut}{Can be "in" or "out". Condition of searching. "in" means that end of MACS-peak can be after param "end".
"out" means that end of MACS-peak can be only before param "end".}

\item{massTF}{List of TFs or quantity of MACS-peaks to choose CRE.}

\item{powerNR}{Minimal needed quantity of MACS-peaks to choose CRE.}

\item{limitPart}{Minimal needed fraction of TFs.}
}
\value{
Data Frame (class = "data.frame") of CREs, dominante cellular line, quantity of type of cellular lines,
quantity of different TFs, quantity of TFs of dominating cellular line
and fraction of TFs of dominating cellular line.
}
\description{
Function counts fraction of factors of the dominating cellular line.
}
\examples{
myFrame <- fractionCellTF(arrayData = "", nameChromosome = "chr21", begin = "", end = "",
firstCut = "in", secondCut = "in", massTF = 50, powerNR = 3, limitPart = 0.2)

}
\author{
Alexey Solovyev
}
