% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_TF.R
\name{frequencyTF}
\alias{frequencyTF}
\title{How many regions contain the factor}
\usage{
frequencyTF(arrayData = data.frame)
}
\arguments{
\item{arrayData}{Data from one of Fetches.}
}
\value{
The sorted list of TFs and the quantity of regions.
}
\description{
Function counts the quantity of regions where TF can be found.
}
\examples{
myList <- frequencyTF(arrayData = "")

}
\author{
Alexey Solovyev
}
