% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_anonyme.R
\name{hiAnonyme}
\alias{hiAnonyme}
\title{Verification of loading of package}
\usage{
hiAnonyme(who)
}
\arguments{
\item{who}{The name of the person you wish to start a
conversation with.}
}
\value{
A line of text to be used when starting conversations.
}
\description{
Help start conversations. This function generates some helpful text that can be used to start
conversations in all kinds of awkward social situations.
}
\examples{
hi("Alexey Solovyev")

}
\author{
Alexey Solovyev
}
