% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventarisation.R
\name{inventarisation}
\alias{inventarisation}
\title{Path to choosen set}
\usage{
inventarisation(mode_res = "", version = "", species = "", assembly = "", type = "")
}
\arguments{
\item{mode_res}{Determines the type of answer, it can be
'path', 'version', 'species', 'assembly' or 'type'.}

\item{version}{Version of ReMap.}

\item{species}{Name of specie.}

\item{assembly}{Number of assembly.}

\item{type}{Type of data.}
}
\value{
Path to the chosen file.
}
\description{
Function gives possibility to select the set of data.
}
\examples{
remapFile <- inventarisation(mode_res = "path", version = "2018",
species = "Homo sapiens", assembly = "hg38", type = "All peaks")

which_version <- inventarisation(mode = "version")

}
\author{
Alexey Solovyev
}
