% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_to_tmp.R
\name{loadFromUrlToTmp}
\alias{loadFromUrlToTmp}
\title{Load file from URL to tmp}
\usage{
loadChromFile <- function(genome)
}
\arguments{
\item{genome}{The name of the species to import the chromosomes from.}
}
\value{
The path to a file that contains the chromosome lengths.
}
\description{
Load files.
}
\examples{
hg19ChromFile <- loadChromFile("hg19")

}
\author{
Alexey Solovyev
}
