% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_annexe.R
\name{makeAnnexe}
\alias{makeAnnexe}
\title{Counts MACS-peaks on chromosome}
\usage{
makeAnnexe(dirFrom)
}
\arguments{
\item{dirFrom}{The path to output annexe-file. By default "~/tmp/RData/Output".}
}
\value{
Table of quantity of MACS-peaks on chromosome. In case of success it returns the word "Done".
}
\description{
Function counts MACS-peaks on chromosome.
}
\examples{
makeAnnexe(dirFrom="")

}
\author{
Alexey Solovyev
}
