% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_series.R
\name{makeSummarySeries}
\alias{makeSummarySeries}
\title{Counts some total information}
\usage{
makeSummarySeries(dirFrom)
}
\arguments{
\item{dirFrom}{The path to output file. By default "~/tmp/RData/Output".}
}
\value{
Table of quantity of sum of peaks, TFs and cellular lines for all IDs.
In case of success it returns the word "Done".
}
\description{
Function counts sum of peaks, TFs and cellular lines for all IDs.
}
\examples{
makeSummarySeries(dirFrom="")

}
\author{
Alexey Solovyev
}
