% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range.R
\name{plotRange}
\alias{plotRange}
\title{Test unused function}
\usage{
plotRange(x, xlim = x, main = deparse(substitute(x)), col = "black",
  sep = 0.5, ...)
}
\description{
Function plots IRange.
}
\examples{
hg19ChromFile(x = IRagne)

}
\author{
Alexey Solovyev
}
