% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBS.R
\name{TFBS}
\alias{TFBS}
\title{CREs for given TF}
\usage{
TFBS(list of params)
}
\arguments{
\item{arrayData}{Data or path to data, by default "~/tmp/RData/Output".}

\item{nameChromosome}{Name of chromosome.}

\item{begin}{Position of starting of searching on chromosome.}

\item{end}{Position of ending of searching on chromosome.}

\item{firstCut}{Can be "in" or "out". Condition of searching. "in" means that begin of MACS-peak can be before param "begin".
"out" means that begin of MACS-peak can be only after param "begin".}

\item{secondCut}{Can be "in" or "out". Condition of searching. "in" means that end of MACS-peak can be after param "end".
"out" means that end of MACS-peak can be only before param "end".}

\item{TF}{Name of TF.}

\item{powerNR}{Minimal needed quantity of MACS-peaks to choose CRE.}
}
\value{
Data Frame (class = "data.frame") of set of CREs.
}
\description{
Function searches the CREs for given TF in region.
}
\examples{
myFrame <- TFBS(arrayData = "", nameChromosome = "chr21", begin = 5240000, end = 52450000,
firstCut = "in", secondCut = "in", TF = "AR", powerNR = 3)

}
\author{
Alexey Solovyev
}
