% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_RData.R
\name{bedToRData}
\alias{bedToRData}
\title{Load chromosomes sizes file}
\usage{
loadChromFile <- function(genome)
}
\arguments{
\item{genome}{The name of the species to import the chromosomes from.}
}
\value{
The path to a file that contains the chromosome lengths.
}
\description{
Load preloaded chromosome sizes files.
}
\examples{
hg19ChromFile <- loadChromFile("hg19")

}
\author{
Alexey Solovyev
}
