% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_to_tmp.R
\name{loadFromUrlToTmp}
\alias{loadFromUrlToTmp}
\title{Loads file of listing remapR from URL to ~/tmp}
\usage{
loadFromUrlToTmp(url.in, tmp.to)
}
\arguments{
\item{url.in}{The path to a listing-file.}

\item{tmp.to}{The name of listing-file.}
}
\value{
Data.Frame from the listing-file.
}
\description{
Loads listing-file and creates ~/tmp.
}
\examples{
myFrame <- loadFromUrlToTmp("http://tagc.univ-mrs.fr/remap/download/remapR", "remapR.csv")

}
\author{
Alexey Solovyev
}
